requirements.txt
yt-dlp
moviepy
openai-whisper
googletrans==4.0.0-rc1
torch
torchaudio
transformers
subprocess.run
ffmpeg
pydub
parler_tts
flask-cors
secure-smtplib
requests
faster_whisper
flask==3.0.2
gunicorn==21.2.0
pymongo==4.6.1
python-dotenv==1.0.0